+++ ovp_core/serializers.py
@@ -7,3 +7,13 @@ class GoogleAddressSerializer(serializers.ModelSerializer):
     model = models.GoogleAddress
     fields = ['typed_address', 'typed_address2', 'address_line', 'city_state']
     read_only_fields = ['address_line', 'city_state']
+
+class SkillSerializer(serializers.ModelSerializer):
+  class Meta:
+    fields = ['name']
+    model = models.Skill
+
+class CauseSerializer(serializers.ModelSerializer):
+  class Meta:
+    fields = ['name']
+    model = models.Cause